requirements.txt
python-telegram-bot==21.0
yt-dlp
mutagen
requests